requirements.txt
python-docx==0.8.11
sentence-transformers==2.2.2
transformers==4.30.2
torch==1.13.1
faiss-cpu==1.7.4
numpy==1.23.5
huggingface_hub==0.10.1